requirements.txt
python-dotenv
Faker
psycopg2-binary
selenium